# docs/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: mandarine.app@yandex.ru
    name: Mandarine Support
  description: API for web and mobile application Mandarine
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Mandarine API
  version: 0.0.0
paths:
...
  "/v0/resources/{objectID}":
    get:
      description: Request for getting resource. Return the resource in S3 storage.
      operationId: DownloadResource
      parameters:
        - description: Object id
          in: path
          name: objectID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      summary: Download resource
      tags:
        - Resource API
  /v0/resources/many:
    post:
      description: Request for uploading resources. Return the array of object ids in
        S3 storage for successful uploaded files.
      operationId: UploadResources
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resources:
                  description: Files to upload
                  type: array
                  items:
                    type: string
              required:
                - resources
        required: true
      responses:
        "201":
          description: Uploaded resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.UploadResourcesOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Upload resources
      tags:
        - Resource API
  /v0/resources/one:
    post:
      description: Request for uploading resource. Return the object id in S3 storage.
      operationId: UploadResource
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resource:
                  description: File to upload
                  type: string
              required:
                - resource
        required: true
      responses:
        "201":
          description: Uploaded resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.UploadResourceOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Upload resource
      tags:
        - Resource API
  /v0/ws:
    get:
      description: Request for connect to websocket server. If pool is not full, a new
        websocket connection is created.
      operationId: WsConnect
      responses:
        "101":
          description: Switching Protocols
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Connect to websocket server
      tags:
        - Websocket API
...
servers:
  - url: //localhost:8080
components:
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    v0.ErrorOutput:
      properties:
        message:
          type: string
        path:
          example: /api/v0/example
          format: url_path
          type: string
        status:
          maximum: 599
          minimum: 400
          type: integer
        timestamp:
          type: string
      required:
        - message
        - path
        - status
        - timestamp
      type: object
...
    v0.UploadResourceOutput:
      properties:
        object_id:
          type: string
      type: object
    v0.UploadResourcesOutput:
      properties:
        count:
          type: integer
        data:
          additionalProperties:
            $ref: "#/components/schemas/v0.UploadResourceOutput"
          type: object
      type: object
...